+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 > Patch
 - **commitlint:** added patch support
+- **fzf:** added live grep current buffer `<c-y>`
 
 
 <a name="v4.4.1"></a>