+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - **fzf:** added live grep current buffer `<c-y>`
 - **neotree:** Neotree Reveal map now uses vim.keymap.set
 
+> Various Actions
+- update snapshot
+
 
 <a name="v4.4.1"></a>
 ### [v4.4.1] - 2022-05-18